Add Pug (formerly Jade) support, start /clock
Pug is an HTML templating engine which allows you to write HTML in a more efficient and cleaner way. Much like Sass is a preprocessor for CSS, Pug accomplishes that for HTML. The clock page was created as a means of testing out the structure of a tournament clock. Will need to add better support for stylesheets (using .scss files) and clean-up code with extra whitespace/poor formatting.